# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Students LPNU",
    "description": "Lab 7 ",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/groups": {
      "post": {
        "summary": "Add new student to a group",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        },
        "tags": [
          "Group"
        ],
        "operationId": "AddStudent",
        "responses": {
          "201": {
            "description": "Successfully added student",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/teacher": {
      "post": {
        "summary": "Add mark for a student",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mark"
              }
            }
          }
        },
        "tags": [
          "Teacher"
        ],
        "operationId": "MarkStudent",
        "responses": {
          "201": {
            "description": "Successfully marked student",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Mark"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/teachers": {
      "post": {
        "summary": "Add teacher",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Teacher"
              }
            }
          }
        },
        "tags": [
          "Faculty"
        ],
        "operationId": "AddTeacher",
        "responses": {
          "201": {
            "description": "Successfully added a teacher",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Teacher"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get All teachers",
        "tags": [
          "Faculty"
        ],
        "operationId": "GetAllTeachers",
        "responses": {
          "200": {
            "description": "Successfully get all teachers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Teacher"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Student": {
        "type": "object",
        "properties": {
          "StudentId": {
            "type": "string",
            "example": "Alaska"
          },
          "Firstname": {
            "type": "string",
            "example": "Kornelia"
          },
          "Surname": {
            "type": "string",
            "example": "Drozd"
          },
          "Password": {
            "type": "string",
            "example": "qwerty1234"
          },
          "GroupId": {
            "type": "integer",
            "minimum": 214,
            "maximum": 217,
            "example": 216
          }
        }
      },
      "Teacher": {
        "type": "object",
        "properties": {
          "TeacherId": {
            "type": "string",
            "example": "petroPu"
          },
          "Password": {
            "type": "string",
            "example": "qwerty"
          },
          "Firstname": {
            "type": "string",
            "example": "Petro"
          },
          "Surname": {
            "type": "string",
            "example": "Pukach"
          },
          "SubjectId": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "Mark": {
        "type": "object",
        "properties": {
          "MarkId": {
            "type": "integer",
            "example": 1
          },
          "StudentId": {
            "type": "string",
            "example": "Alaska"
          },
          "SubjectId": {
            "type": "integer",
            "example": 1
          },
          "TeacherId": {
            "type": "string",
            "example": "petroPu"
          },
          "Date": {
            "type": "string",
            "format": "date",
            "example": "2022-08-01"
          },
          "Value": {
            "type": "number",
            "format": "float",
            "example": 4.5
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}